Refactor(codebase): css library
Removed MUI and replaced with TailwindCSS
