+++ .prettierrc.cjs
@@ -1,7 +1,8 @@
 // prettier.config.js or .prettierrc.js
 module.exports = {
-    trailingComma: "es5",
-    tabWidth: 2,
-    semi: true,
-    singleQuote: false,
-  };
+  trailingComma: "es5",
+  tabWidth: 2,
+  semi: true,
+  singleQuote: false,
+  plugins: [require("prettier-plugin-tailwindcss")],
+};
