+++ src/App.jsx
@@ -1,26 +1,17 @@
-import { CssBaseline } from "@mui/material";
-import { ThemeProvider ,createTheme } from "@mui/material";
 import "./App.css";
 import { Routes, Route } from "react-router-dom";
 
 // Import Pages
 import Admin from "./pages/Admin/AdminDashboard";
-import Login from "./pages/Login/Login";
-import Register from "./pages/Register";
+import Login from "./pages/Auth/Login/";
+import Register from "./pages/Auth/Register";
 import Transporter from "./pages/Transporter/TransporterDashboard";
 import Client from "./pages/Client/ClientDashboard";
 
 function App() {
-  const theme = createTheme({
-    typography: {
-      fontFamily: [
-        'inter',
-      ].join(','),
-    },});
   return (
-    <ThemeProvider theme={theme}>
+
     <div className="App">
-      <CssBaseline />
       <Routes>
         <Route path="/login" element={<Login />} />
         <Route path="/register" element={<Register />} />
@@ -29,8 +20,7 @@ function App() {
         <Route path="/transporter" element={<Transporter />} />
         <Route path="*" element={<h1>404: Not found </h1>} />
       </Routes>
-      </div>
-    </ThemeProvider>
+    </div>
   );
 }
