+++ src/components/TextInput/TextInput.css
@@ -1,7 +1,7 @@
 input:focus-visible{
     -webkit-appearance: none;
     appearance: none;
-    outline: 2px solid #00000000;;
+    outline: 1px solid #00000000;;
 }
 .reTextInput{
     width: 100%;
@@ -18,7 +18,7 @@ input:focus-visible{
 }
 .reTextInput input{
     background: #FFFFFF;
-    border: 2px solid #000000;
+    border: 1px solid #000000;
     box-shadow: 4px 6px 15px rgba(0, 0, 0, 0.06);
     border-radius: 6px;
     padding: 15px;
