+++ src/pages/Auth/Verify/Index.jsx
@@ -1,0 +1,5 @@
+const Verify = () => {
+    
+}
+
+export default Verify
